requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
numpy
pandas>=2.0
pyarrow
pydantic>=2.0
pyyaml
python-dotenv

# ── LLM providers ───────────────────────────────────────────────────────────
anthropic>=0.40.0
google-genai>=1.0.0

# ── Dataset (HuggingFace) ───────────────────────────────────────────────────
datasets

# ── Analysis ─────────────────────────────────────────────────────────────────
scikit-learn

# ── Visualization ────────────────────────────────────────────────────────────
matplotlib>=3.7
seaborn>=0.13
